mimic java file structure
